requirements.txt
django
djangorestframework
django-redis
pytest
pytest-django
model-bakery

gunicorn
psycopg2-binary
Pillow
django-cors-headers
